requirements.txt
# Deep Learning Acceleration (Removing extra-index to favor Conda-installed Torch)
xformers==0.0.29
accelerate>=1.1.0

# LLM & Data Processing
transformers>=4.46.0
datasets>=3.1.0
einops==0.8.2
tqdm>=4.67.1

# 3D Vision & Graphics (Foundational for PyTorch3D)
iopath
fvcore
open3d>=0.18.0
mathutils>=3.3.0
torchdiffeq==0.2.5

# Image Processing & Rendering
imageio>=2.33.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
imageio[ffmpeg]
imageio[pyav]
# Scientific Computing (Keeping NumPy < 2.0 is smart for PyTorch 2.5 stability)
numpy<2.1.0
matplotlib>=3.8.0
PyYAML>=6.0.3
omegaconf>=2.3.0
torchmetrics>=1.5.0

# Utilities
packaging>=26.0
setuptools>=75.0.0
tensorboard